refactor project structure in more
