+++ IronRoadForChildren/Features/Map/PhotoDetailView.swift
@@ -4,29 +4,28 @@ import PDFKit
 import SwiftUI
 
 class CustomPDFView: PDFView {
-    private var initialScaleFactor: CGFloat?
+	private var initialScaleFactor: CGFloat?
 
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        if initialScaleFactor == nil {
-            initialScaleFactor = self.scaleFactor
-            self.minScaleFactor = initialScaleFactor!
-        }
-    }
+	override func layoutSubviews() {
+		super.layoutSubviews()
+		if initialScaleFactor == nil {
+			initialScaleFactor = scaleFactor
+			minScaleFactor = initialScaleFactor!
+		}
+	}
 }
 
 struct PhotoDetailView: UIViewRepresentable {
-    let image: UIImage
+	let image: UIImage
 
-    func makeUIView(context: Context) -> PDFView {
-        let view = CustomPDFView()
-        view.document = PDFDocument()
-        guard let page = PDFPage(image: image) else { return view }
-        view.document?.insert(page, at: 0)
-        view.autoScales = true
-        return view
-    }
+	func makeUIView(context _: Context) -> PDFView {
+		let view = CustomPDFView()
+		view.document = PDFDocument()
+		guard let page = PDFPage(image: image) else { return view }
+		view.document?.insert(page, at: 0)
+		view.autoScales = true
+		return view
+	}
 
-    func updateUIView(_ uiView: PDFView, context: Context) {
-    }
+	func updateUIView(_: PDFView, context _: Context) {}
 }
